+++ resources/views/roles/list.blade.php
@@ -40,13 +40,17 @@
                                     {{ \Carbon\Carbon::parse($role->created_at)->format('d M, Y') }}
                                 </td>
                                 <td class="px-6 py-3 text-center">
-                                    <a href="{{ route('roles.edit', $role->id) }}"
-                                        class="bg-slate-700 text-sm rounded-md px-3 py-2 text-white hover:bg-slate-600">Edit
-                                    </a>
+                                    @can('edit roles')
+                                        <a href="{{ route('roles.edit', $role->id) }}"
+                                            class="bg-slate-700 text-sm rounded-md px-3 py-2 text-white hover:bg-slate-600">Edit
+                                        </a>
+                                    @endcan
 
-                                    <a href="javascript:void(0);" onclick="deleteRole({{$role->id}})"
-                                        class="bg-red-700 text-sm rounded-md px-3 py-2 text-white hover:bg-red-600">Delete
-                                    </a>
+                                    @can('delete roles')
+                                        <a href="javascript:void(0);" onclick="deleteRole({{ $role->id }})"
+                                            class="bg-red-700 text-sm rounded-md px-3 py-2 text-white hover:bg-red-600">Delete
+                                        </a>
+                                    @endcan
                                 </td>
                             </tr>
                         @endforeach
@@ -62,17 +66,19 @@
 
     <x-slot name="script">
         <script type="text/javascript">
-            function deleteRole(id){
-                if(confirm("Are you sure you want to delete?")){
+            function deleteRole(id) {
+                if (confirm("Are you sure you want to delete?")) {
                     $.ajax({
-                        url:'{{ route("roles.destroy") }}',
-                        type:'delete',
-                        data: {id:id},
+                        url: '{{ route('roles.destroy') }}',
+                        type: 'delete',
+                        data: {
+                            id: id
+                        },
                         dataType: 'json',
                         headers: {
-                            'x-csrf-token' : '{{ csrf_token() }}'
+                            'x-csrf-token': '{{ csrf_token() }}'
                         },
-                        success: function(response){
+                        success: function(response) {
                             window.location.href = '{{ route('roles.index') }}'
                         }
                     })
